+++ source/acquaman/SGM/SGMXASScanConfiguration.h
@@ -38,7 +38,7 @@ public:
 	QList<AMDetectorInfo*> usingDetectors() const;
 
 public slots:
-	virtual bool addRegion(size_t index, double start, double delta, double end) { return regions_->addRegion(index, start, delta, end);}
+	virtual bool addRegion(int index, double start, double delta, double end) { return regions_->addRegion(index, start, delta, end);}
 
 	bool setExitSlitGap(double exitSlitGap);
 	bool setGrating(SGMBeamline::sgmGrating grating);
