+++ source/acquaman/dacq3_2/xmlWrite.cpp
@@ -294,7 +294,7 @@ ConvertInput(const char *in, const char *encoding)
 
     size = (int) strlen(in) + 1;
     out_size = size * 2 - 1;
-    out = (unsigned char *) xmlMalloc((size_t) out_size);
+    out = (unsigned char *) xmlMalloc((int) out_size);
 
     if (out != 0) {
         temp = size - 1;
